Renamed amenities bar to top bar to be more suitable, moved show listings and display listings buttons to top bar, restyled everything.
