+++ taui/src/sass/01_settings/_color.scss
@@ -5,6 +5,7 @@ $gray-200: #eee;
 $gray-300: #e0e0e0;
 $gray-400: #bdbdbd;
 $gray-500: #9e9e9e;
+$gray-550: #9aa0a6;
 $gray-600: #757575;
 $gray-700: #616161;
 $gray-800: #424242;
