+++ taui/src/sass/06_components/_top-bar.scss
@@ -1,0 +1,152 @@
+.top-bar {
+	justify-content: space-between;
+	width: 100%;
+	white-space: nowrap;
+	margin-bottom: 2rem;
+
+	&__bar {
+		display: inline-block;
+	}
+
+	&__heading {
+		@include text(200);
+		@include font-weight(bold);
+		display: inline-block;
+		margin: 0 0.25rem 0 0.5rem;
+	}
+
+	&__button {
+		display: inline-block;
+	    margin: 0 0.25rem;
+	    padding: 0.5rem;
+	    color: $white;
+	    text-align: center;
+	    text-transform: none;
+	    text-decoration: none;
+	    background: $gray-550;
+	    border-radius: 1px;
+	    outline: none;
+	    transition: background 250ms ease-in-out, transform 150ms ease;
+
+	    &::-moz-focus-inner {
+	        border: none;
+	    }
+
+	    &:focus {
+	        outline: 1px solid $white;
+	        outline-offset: -3px;
+	    }
+
+	    &:active {
+	        transform: scale(0.96);
+	    }
+
+	    &:hover:not(:active):not([disabled]) {
+	        transform: scale(1.04);
+	    }
+
+	    &[disabled] {
+	        cursor: not-allowed;
+	        filter: saturate(12%);
+	    }
+	}
+
+	&__button-highlighted {
+		display: inline-block;
+	    margin: 0 0.25rem;
+	    padding: 0.5rem;
+	    color: $white;
+	    text-align: center;
+	    text-transform: none;
+	    text-decoration: none;
+	    background: $primary;
+	    border-radius: 1px;
+	    outline: none;
+	    transition: background 250ms ease-in-out, transform 150ms ease;
+
+	    &::-moz-focus-inner {
+	        border: none;
+	    }
+
+	    &:focus {
+	        outline: 1px solid $white;
+	        outline-offset: -3px;
+	    }
+
+	    &:active {
+	        transform: scale(0.96);
+	    }
+
+	    &:hover:not(:active):not([disabled]) {
+	        transform: scale(1.04);
+	    }
+
+	    &[disabled] {
+	        cursor: not-allowed;
+	        filter: saturate(12%);
+	    }
+	}
+
+	&__button-amenities {
+		display: inline-block;
+	    margin: 0 0.25rem;
+	    padding: 0 0.5rem 0.5rem 0.5rem;
+	    color: $white;
+	    text-align: center;
+	    text-transform: none;
+	    text-decoration: none;
+	    background: $gray-550;
+	    border-radius: 1px;
+	    outline: none;
+	    transition: background 250ms ease-in-out, transform 150ms ease;
+
+	    &::-moz-focus-inner {
+	        border: none;
+	    }
+
+	    &:focus {
+	        outline: 1px solid $white;
+	        outline-offset: -3px;
+	    }
+
+	    &:active {
+	        transform: scale(0.96);
+	    }
+
+	    &:hover:not(:active):not([disabled]) {
+	        transform: scale(1.04);
+	    }
+
+	    &[disabled] {
+	        cursor: not-allowed;
+	        filter: saturate(12%);
+	    }
+	}
+
+	&__amenities-all {
+	  z-index: 1000;
+	  display: inline-block;
+	  white-space: no-wrap;
+	  align-items: center;
+	  padding: 0% 2%;
+	  @include text(200);
+	  @include font-weight(bold);
+	}
+
+	// &__amenities-button {
+	// 	@include button($color: $white, $background: $gray-550);
+	// 	margin: 0 0.1rem;
+	// 	display: table-cell;
+	// 	align-items: center;
+	// }
+
+
+	&__icon {
+	  position: relative;
+	  display: inline-block;
+	  top: 0.5rem;
+	  width: 2rem;
+	  z-index: 99;
+	}
+
+}
