+++ taui/src/sass/main.scss
@@ -23,6 +23,7 @@
 '06_components/neighborhood-details.scss',
 '06_components/rental-units-meter.scss',
 '06_components/text-alert-popup.scss',
+'06_components/top-bar.scss',
 '07_layouts/app-root.scss',
 '07_layouts/auth-screen.scss',
 '07_layouts/form-screen.scss',
